requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.28
Flask-WTF==1.0.0
Flask-Bootstrap==3.3.7.1
boto3==1.18.64